pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentos-mvp"
version = "0.1.0"
description = "AgentOS MVP (local/cloud)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.0",
  "PyYAML>=6.0",
  "httpx>=0.27",
  "tenacity>=8.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
include = ["agentos*"]
exclude = ["config*", "deployments*", "tests*", "docs*", "scripts*"]
